Ejercicio 4

+++ Programming/Unit2/Herencia/Ejercicio4/Secciones/Objetos/Dimensiones.java
@@ -2,27 +2,27 @@ package Programming.Unit2.Herencia.Ejercicio4.Secciones.Objetos;
 
 public class Dimensiones {
 
-    private int altura;
-    private int anchura;
+    private double altura;
+    private double anchura;
 
-    public Dimensiones(int altura, int anchura) {
+    public Dimensiones(double altura, double anchura) {
         this.altura = altura;
         this.anchura = anchura;
     }
 
-    public int getAltura() {
+    public double getAltura() {
         return altura;
     }
 
-    public void setAltura(int altura) {
+    public void setAltura(double altura) {
         this.altura = altura;
     }
 
-    public int getAnchura() {
+    public double getAnchura() {
         return anchura;
     }
 
-    public void setAnchura(int anchura) {
+    public void setAnchura(double anchura) {
         this.anchura = anchura;
     }
 }
